Refactoring Video, VideoAI, test_VideoProcessor.py.
