+++ src/pyWitnessAI/__init__.py
@@ -1,16 +1,15 @@
-from .Constants import *
+from pyWitnessAI.utils.Constants import *
 from .Images import *
 from .ImagesAI import *
 
 from .Video import *
-from .VideoAnalyzer import *
+from .VideoAI import *
 from .VideoProcessor import *
 
 from .Lineup import *
 
-from .Test import *
 from .LineupDecider import *
-from .LineupPipeline import *
+from .VideoLineupPipeline import *
 from .utils import find_bins, plot_hist_with_edges
 
 __all__ = ["find_bins", "plot_hist_with_edges"]
